GetComponents: remove use of $orig_dir to support whitespace in it
Using $orig_dir in command expansions become very difficult once
$orig_dir contains whitespace and single quotes eg "Roland's Documents".
There is no real need to do so since it was just path to cwd() anyway.
Another reason not to use it is that cwd() returns the actual path which
can differ from what the shell shows if cwd() is a symbolic link.